fix: move presentation utils to utils foler
