+++ src/components/Canvas/CanvasMain.vue
@@ -52,8 +52,8 @@ import { useAxesStore } from '@/store/axes'
 import { useCanvasStore } from '@/store/canvas'
 import { useDatasetsStore } from '@/store/datasets'
 import { mapState, mapActions } from 'pinia'
-import { getMouseCoordFromMouseEvent } from '@/presentation/mouseEventUtilities'
-import { getRectCoordsFromDragCoords } from '@/presentation/dragRectangleCalculator'
+import { getMouseCoordFromMouseEvent } from '@/presentation/utils/mouseEventUtilities'
+import { getRectCoordsFromDragCoords } from '@/presentation/utils/dragRectangleCalculator'
 
 import { Interpolator } from '@/application/services/interpolator'
 import { HTMLCanvas } from '@/presentation/dom/HTMLCanvas'
