+++ src/presentation/utils/dragRectangleCalculator.test.ts
