+++ src/presentation/utils/dragRectangleCalculator.ts
