+++ src/presentation/utils/mouseEventUtilities.test.ts
