+++ src/presentation/utils/mouseEventUtilities.ts
